+++ src/commands/workspace.rs
@@ -15,6 +15,9 @@ pub enum Commands {
   Delete {
     name: String
   },
+  Cd {
+    name: String
+  },
   Which {
     name: String
   }
@@ -30,6 +33,9 @@ pub fn run(cmd: Commands) -> Result<(), UtilityError> {
     },
     Commands::Delete { name } => {
       cmd_delete(name)
+    },
+    Commands::Cd { name } => {
+      cmd_cd(name)
     },
     Commands::Which { name } => {
       cmd_which(name)
@@ -103,7 +109,8 @@ fn cmd_init(name: String) -> Result<(), UtilityError> {
       version: env!("CARGO_PKG_VERSION").to_string(),
     },
     workspace: Some(HashMap::new()),
-    repository: None
+    repository: None,
+    scripts: None
   };
 
   write_toml(&batl_toml_path, &config)?;
@@ -135,6 +142,32 @@ fn cmd_delete(name: String) -> Result<(), UtilityError> {
   std::fs::remove_dir_all(path)?;
 
   success(&format!("Workspace {} deleted", name));
+
+  Ok(())
+}
+
+fn cmd_cd(name: String) -> Result<(), UtilityError> {
+  if !BATL_NAME_REGEX.is_match(&name) {
+    return Err(UtilityError::InvalidName(name));
+  }
+
+  let workspace_root = get_batl_root()?.join("workspaces");
+
+  let parts = name.split('/').collect::<Vec<&str>>();
+  let mut path = workspace_root;
+
+  for part in parts.iter().take(parts.len() - 1) {
+    path = path.join(format!("@{}", part));
+  }
+  path = path.join(parts.last().unwrap());
+
+  if !path.exists() {
+    return Err(UtilityError::ResourceDoesNotExist(format!("Workspace {}", name)));
+  }
+
+  std::env::set_current_dir(path)?;
+
+  success(&format!("Workspace {} selected", name));
 
   Ok(())
 }