+++ src/config.rs
@@ -1,11 +1,15 @@
 use serde::{Serialize, Deserialize};
 use std::collections::HashMap;
 
+/****************************************
+* Config Formats
+****************************************/
 #[derive(Serialize, Deserialize, Clone)]
 pub struct Config {
   pub environment: EnvConfig,
   pub workspace: Option<HashMap<String, String>>,
-  pub repository: Option<RepositoryConfig>
+  pub repository: Option<RepositoryConfig>,
+  pub scripts: Option<HashMap<String, String>>
 }
 
 #[derive(Serialize, Deserialize, Clone)]
@@ -17,5 +21,5 @@ pub struct EnvConfig {
 pub struct RepositoryConfig {
   pub name: String,
   pub version: String,
-  pub build: String
+  pub build: Option<String>
 }